# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateIrrigationCommandRequest:
    properties:
      action:
        enum:
        - START
        - STOP
        type: string
      device_id:
        type: string
      duration:
        type: integer
    required:
    - action
    - device_id
    type: object
  dto.CreateIrrigationCommandResponse:
    properties:
      command_id:
        type: string
      status:
        type: string
    type: object
...
  dto.FaqResponse:
    properties:
      answer:
        type: string
      category:
        type: string
      order_number:
        type: integer
      question:
        type: string
      uuid:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: contato@groundguard.com
    name: Guilherme Teixeira
  description: |-
    API REST do Ground Guard, uma plataforma IoT para monitoramento e automação de jardins e plantas.
    Permite gerenciamento de dispositivos, preferências de irrigação, monitoramento ambiental e acionamento remoto de irrigação.
    Desenvolvido como TCC e preparado para evolução comercial.
  license:
    name: MIT
  termsOfService: https://groundguard.com/terms
  title: Ground Guard API
  version: 1.0.0
paths:
...
  /faqs:
    get:
      consumes:
      - application/json
      description: Retorna uma lista paginada de FAQs
      parameters:
      - description: Número da página
        in: query
        minimum: 1
        name: page_id
        required: true
        type: integer
      - description: Tamanho da página
        in: query
        maximum: 100
        minimum: 1
        name: page_size
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Sucesso
          schema:
            items:
              $ref: '#/definitions/dto.FaqResponse'
            type: array
        "400":
          description: Requisição inválida (Erro de paginação)
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Erro interno do servidor
          schema:
            additionalProperties: true
            type: object
      summary: Listar FAQs
      tags:
      - faqs
  /faqs/{uuid}:
    get:
      consumes:
      - application/json
      description: Retorna os detalhes de um FAQ específico baseado no seu UUID
      parameters:
      - description: UUID do FAQ
        format: uuid
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Sucesso
          schema:
            $ref: '#/definitions/dto.FaqResponse'
        "400":
          description: Requisição inválida (Erro de bind ou UUID mal formatado)
          schema:
            additionalProperties: true
            type: object
        "404":
          description: FAQ não encontrado
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Erro interno do servidor
          schema:
            additionalProperties: true
            type: object
      summary: Obter um FAQ
      tags:
      - faqs
...
  /irrigation/command/{uuid}:
    get:
      consumes:
      - application/json
      description: Retorna os detalhes de um comando de irrigação específico através
        do seu UUID.
      parameters:
      - description: UUID do comando de irrigação
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CreateIrrigationCommandResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Obter comando de irrigação
      tags:
      - irrigation
  /irrigation/commands:
    post:
      consumes:
      - application/json
      description: Registra um comando pendente para ser enviado ao ESP32 via MQTT.
      parameters:
      - description: Dados do comando
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateIrrigationCommandRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CreateIrrigationCommandResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Criar comando de irrigação
      tags:
      - irrigation
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Informe: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"